+++ hfut/__init__.py
@@ -11,7 +11,7 @@ from .shortcut import *
 from .value import *
 
 __title__ = 'hfut'
-__version__ = '2.0.0'
+__version__ = '2.1.0'
 __author__ = 'erliang'
 __author_email__ = 'dev@erliang.me'
 __url__ = 'https://github.com/er1iang/hfut'