+++ tests/conftest.py
@@ -1,5 +1,6 @@
 # -*- coding:utf-8 -*-
 """
+http://docs.pytest.org/en/latest/fixture.html
 http://docs.pytest.org/en/latest/writing_plugins.html#conftest-py-local-per-directory-plugins
 """
 from __future__ import unicode_literals
@@ -9,17 +10,22 @@ import os
 
 import pytest
 
-from hfut import Student
+from hfut import Student, ENV
 
 base_dir = os.path.dirname(__file__)
 
 
 def load_shortcuts():
     with open(os.path.join(base_dir, 'test_accounts.csv')) as fp:
-        shortcuts = [Student(account, password, campus) for account, password, campus in csv.reader(fp)]
-    return shortcuts
+        for account, password, campus in csv.reader(fp):
+            yield Student(account, password, campus)
 
 
 @pytest.fixture(params=load_shortcuts())
 def shortcuts(request):
     return request.param
+
+
+@pytest.fixture(scope='class', params=['html.parser', 'lxml'])
+def features(request):
+    ENV['SOUP_FEATURES'] = request.param